+++ client/src/app/create/create.component.ts
@@ -135,20 +135,6 @@ export class CreateComponent implements OnInit, AfterViewInit, OnDestroy {
       this.weekForm.value).subscribe(
         workout => {
           MaterialService.toast('Added new workout!');
-
-          for (const days of this.weekForm.value.weeks) {
-            for (const exercises of days.days) {
-              for (const exercise of exercises.exercises) {
-                this._baseService.updateUse(exercise.exercise).subscribe(
-                  base => { },
-                  error => {
-                    MaterialService.toast(error.error.message);
-                  }
-                );
-              }
-            }
-          }
-
           this.initForm.reset();
           this.initForm.enable();
           this.weekForm.disable();