+++ routes/base.js
@@ -10,7 +10,6 @@ router.post('/', passport.authenticate('jwt', { session: false }), upload.single
 router.get('/', passport.authenticate('jwt', { session: false }), controller.getAllExercises);
 router.get('/:id', passport.authenticate('jwt', { session: false }), controller.getExercise);
 router.delete('/:id', passport.authenticate('jwt', { session: false }), controller.deleteExercise);
-router.patch('/', passport.authenticate('jwt', { session: false }), controller.updateUse);
 router.patch('/:id', passport.authenticate('jwt', { session: false }), upload.single('video'), controller.update);
 
 module.exports = router;